+++ changelog.md
@@ -20,6 +20,7 @@
   - Don’t specify the `CI` environment variable as it’s set by default.
 - Simplified the ESLint config.
 - Removed `npm-debug.log` from the `.gitignore` file as npm [v4.2.0](https://github.com/npm/npm/releases/tag/v4.2.0)+ doesn’t create it in the current working directory.
+- Removed the test helper function `sleep`.
 
 ## 3.0.0